Cast one of the operands to double, to make it clear, that we are using the result as double.
git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1783505 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: b04c8f79f0cb9fd2b5a38185d0fe9f203388bcb1

+++ src/core/org/apache/jmeter/report/processor/graph/impl/ResponseTimeDistributionGraphConsumer.java
@@ -72,8 +72,7 @@ public class ResponseTimeDistributionGraphConsumer extends
             @Override
             public Double select(Sample sample) {
                 long elapsed = sample.getElapsedTime();
-                return Double.valueOf(elapsed - elapsed % granularity);
-
+                return Double.valueOf((double) elapsed - elapsed % granularity);
             }
         };
     }
